+++ tests/test_v1.py
@@ -4,6 +4,7 @@ import pytest
 
 import pyseto
 from pyseto import DecryptError, EncryptError, Key, VerifyError
+from pyseto.versions.v1 import V1Local, V1Public
 
 from .utils import load_key
 
@@ -58,6 +59,22 @@ class TestV1Local:
             pytest.fail("pyseto.encode() should fail.")
         assert "nonce must be 32 bytes long." in str(err.value)
 
+    @pytest.mark.parametrize(
+        "paserk, msg",
+        [
+            ("xx.local.AAAAAAAAAAAAAAAA", "Invalid PASERK version for a v1.local key."),
+            ("k2.local.AAAAAAAAAAAAAAAA", "Invalid PASERK version for a v1.local key."),
+            ("k1.xxx.AAAAAAAAAAAAAAAA", "Invalid PASERK type for a v1.local key."),
+            ("k1.public.AAAAAAAAAAAAAAAA", "Invalid PASERK type for a v1.local key."),
+        ],
+    )
+    def test_vl_local_from_paserk_with_invalid_args(self, paserk, msg):
+
+        with pytest.raises(ValueError) as err:
+            V1Local.from_paserk(paserk)
+            pytest.fail("Key.from_paserk should fail.")
+        assert msg in str(err.value)
+
 
 class TestV1Public:
     """
@@ -72,3 +89,25 @@ class TestV1Public:
             pyseto.decode(pk, token)
             pytest.fail("pyseto.decode() should fail.")
         assert "Failed to verify." in str(err.value)
+
+    @pytest.mark.parametrize(
+        "paserk, msg",
+        [
+            (
+                "xx.public.AAAAAAAAAAAAAAAA",
+                "Invalid PASERK version for a v1.public key.",
+            ),
+            (
+                "k2.public.AAAAAAAAAAAAAAAA",
+                "Invalid PASERK version for a v1.public key.",
+            ),
+            ("k1.xxx.AAAAAAAAAAAAAAAA", "Invalid PASERK type for a v1.public key."),
+            ("k1.local.AAAAAAAAAAAAAAAA", "Invalid PASERK type for a v1.public key."),
+        ],
+    )
+    def test_vl_public_from_paserk_with_invalid_args(self, paserk, msg):
+
+        with pytest.raises(ValueError) as err:
+            V1Public.from_paserk(paserk)
+            pytest.fail("Key.from_paserk should fail.")
+        assert msg in str(err.value)