+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changes
 Unreleased
 ----------
 
+Version 0.7.0
+-------------
+
+Released 2021-09-16
+
 - Add from_paserk to Key. `#41 <https://github.com/dajiaji/pyseto/pull/41>`__
 - Add support for paserk lid. `#40 <https://github.com/dajiaji/pyseto/pull/40>`__
 - Add support for paserk local. `#40 <https://github.com/dajiaji/pyseto/pull/40>`__